fixed method to find_or_create_by method
method looks for db record with same name and album as passed in as arguments. If it finds it, use that data to create new object with that id, name, album. If it does not find any such record, make new object and save it to db table with #create method. Either way, return song object that was either 'found' or created